Update login, profile and get teacher response
