+++ app/handlers/teachers.go
@@ -7,15 +7,6 @@ import (
 	"github.com/vmkevv/rigelapi/ent/teacher"
 )
 
-type SubWithYear struct {
-	Subscription
-	Year Year `json:"year"`
-}
-type TeacherWithSubs struct {
-	Teacher
-	Subscriptions []SubWithYear `json:"subscriptions"`
-}
-
 func GetTeachers(db *ent.Client) func(*fiber.Ctx) error {
 	return func(c *fiber.Ctx) error {
 		teachers, err := db.Teacher.Query().All(c.Context())
@@ -37,10 +28,6 @@ func GetTeachers(db *ent.Client) func(*fiber.Ctx) error {
 }
 
 func GetTeacher(db *ent.Client) func(*fiber.Ctx) error {
-	type resp struct {
-		Teacher
-		Subscriptions []Subscription `json:"subscriptions"`
-	}
 	return func(c *fiber.Ctx) error {
 		teacherID := c.Params("id")
 		teacher, err := db.Teacher.
@@ -54,30 +41,7 @@ func GetTeacher(db *ent.Client) func(*fiber.Ctx) error {
 		if err != nil {
 			return err
 		}
-		resp := TeacherWithSubs{
-			Teacher: Teacher{
-				ID:       teacherID,
-				Name:     teacher.Name,
-				LastName: teacher.LastName,
-				Email:    teacher.Email,
-				IsAdmin:  teacher.IsAdmin,
-			},
-			Subscriptions: make([]SubWithYear, len(teacher.Edges.Subscriptions)),
-		}
-		for i, subs := range teacher.Edges.Subscriptions {
-			resp.Subscriptions[i] = SubWithYear{
-				Subscription: Subscription{
-					ID:     subs.ID,
-					Method: subs.Method,
-					Qtty:   subs.Qtty,
-					Date:   subs.Date.UnixMilli(),
-				},
-				Year: Year{
-					ID:    subs.Edges.Year.ID,
-					Value: subs.Edges.Year.Value,
-				},
-			}
-		}
+    resp := buildTeacherProfile(teacher)
 		return c.JSON(resp)
 	}
 }
@@ -90,36 +54,14 @@ func GetProfile(db *ent.Client) func(*fiber.Ctx) error {
 			Query().
 			Where(teacher.ID(teacherID)).
 			WithSubscriptions(func(sq *ent.SubscriptionQuery) {
+				sq.WithYear()
 				sq.Order(ent.Asc(subscription.FieldDate))
 			}).
 			First(c.Context())
 		if err != nil {
 			return err
 		}
-		resp := TeacherWithSubs{
-			Teacher: Teacher{
-				ID:       teacherID,
-				Name:     teacher.Name,
-				LastName: teacher.LastName,
-				Email:    teacher.Email,
-				IsAdmin:  teacher.IsAdmin,
-			},
-			Subscriptions: make([]SubWithYear, len(teacher.Edges.Subscriptions)),
-		}
-		for i, subs := range teacher.Edges.Subscriptions {
-			resp.Subscriptions[i] = SubWithYear{
-				Subscription: Subscription{
-					ID:     subs.ID,
-					Method: subs.Method,
-					Qtty:   subs.Qtty,
-					Date:   subs.Date.UnixMilli(),
-				},
-				Year: Year{
-					ID:    subs.Edges.Year.ID,
-					Value: subs.Edges.Year.Value,
-				},
-			}
-		}
+    resp := buildTeacherProfile(teacher)
 		return c.JSON(resp)
 	}
 }
